Added EventMagnitude and IntervalMagnitude classes.
Replaced the drawLabels method in TimelineComponent with drawLabels
method in TimelineTypeInterface.
